Fix: avgScore 타입 변화

+++ src/main/java/com/project/Project/domain/building/BuildingToReviewCategory.java
@@ -20,5 +20,5 @@ public class BuildingToReviewCategory extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     private ReviewCategory reviewCategory;
 
-    private BigDecimal avgScore;
+    private Double avgScore;
 }
